pyproject.toml
[tool.poetry]
name = "repl_python3_theNumberGame-1"
version = "0.1.0"
description = ""
authors = ["ZacKennedy-Smit <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
termcolor = "^1.1.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
